panic for a slice of structs rather than an array of structs
Unpacking into an array of structs would cause a panic. Although it
could unpack into a slice of structs (when there's an linked Sizeof
field) and even an array of built-in types, it could not unpack into
an array of structs. This commit is a quick fix to make it work with
an array of structs while breaking the (unused by us) slice of structs
functionality.